+++ src/components/GoogleLog/googleLogout.tsx
@@ -1,4 +1,5 @@
 import { GoogleLogout } from "react-google-login";
+import { logOut } from "../../util/loginCookies";
 
 const clientId =
   "940221773800-2kpn0pebpu339lmg7tgeu675aooj1mln.apps.googleusercontent.com";
@@ -7,6 +8,7 @@ function Logout() {
 
     const onSuccess = () => {
         alert(' Logout succ');
+        logOut()
     };
   return (
     <div>
