+++ src/components/LoginBox.tsx
@@ -125,7 +125,6 @@ function LoginBox(): JSX.Element {
             <br />
             <div>
               <Login />
-              <Logout />
             </div>
           </LoginFields>
         </form>
